pyproject.toml
[tool.poetry]
name = "Project-DjangoRest"
version = "0.1.0"
description = "Project DjangoRest"
authors = ["Gunther-C <gunther.chev@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
